Work/Code/Python/GRPC/unary_unary/requirements.txt
protobuf
grpcio
pytz